+++ package.json
@@ -4,7 +4,7 @@
 	"version": "1.1.0",
 	"description": "trust_storage is a simple library that allows you to store data in the browser's local storage. It is written in TypeScript and has type definitions.",
 	"main": "dist/index.js",
-	"types": "dist/index.d.ts",
+	"types": "dist/types/index.d.ts",
 	"files": [
 		"dist"
 	],
