+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './models/StorageContext.js';
 export * from './models/GoogleDriveStorage.js';
+export * from './models/CredentialEngine.js';
